+++ project/agent/qlearning/giveFeedback.cc
@@ -1,6 +1,6 @@
 #include "qlearning.ih"
 
-void QLearning::giveFeedback(double reward, size_t newState)
+void QLearning::giveFeedback(float reward, size_t newState)
 {
 //  double &Q_lastAction = d_QTable.at(d_oldstate)[d_lastAction];
 //
