+++ project/agent/qlearning/statespacesize.cc
@@ -5,7 +5,7 @@ void QLearning::stateSpaceSize(size_t size)
 {
   d_QTable.clear();
 
-  vector<double> initialValues { Q_0, Q_0, Q_0, Q_0 };
+  vector<float> initialValues { Q_0, Q_0, Q_0, Q_0 };
   for (size_t idx = 0; idx != size; ++idx)
     d_QTable.push_back(initialValues);
 }
